pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.7"
dependencies = [
    "astrapy>=2.0.1",
    "openai>=1.88.0",
    "pillow>=11.2.1",
    "pymupdf>=1.26.1",
    "pymupdf4llm>=0.0.25",
    "python-dotenv>=1.1.0",
    "requests>=2.32.4",
    "tqdm>=4.67.1",
    "typing-extensions>=4.14.0",
    "voyageai>=0.3.2",
]

[project.optional-dependencies]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0"
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
asyncio_mode = "auto"
